se-backend/pyproject.toml
[project]
name = "se-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.117.1",
    "langgraph>=0.6.7",
    "uvicorn>=0.37.0",
    "openai>=1.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "langchain-community>=0.3.30",
    "colorama>=0.4.6",
    "langchain-core>=0.3.76",
    "dashscope>=1.24.6",
    "langmem>=0.0.29",
    "sqlalchemy>=2.0.0",
    "aiomysql>=0.2.0",
    "pymysql>=1.1.0",
    "greenlet>=3.0.0",
]
